at(index) method is working. This method will return the node that is at the passed index (0-indexed). It works by creating a 'count' variable that will track the number of iterations through each node in the list. Using a while loop we can check to see if the node we're on is null, verifying that it is an actualy node.
Then we evaluate to see if the count is equal to the passed index, if so return the node, else increment the count variable and assign the current node to the next node in the list and restart the while loop.

If the loop runs through without finding the index, it will return the error message.