find(value) method is working. find(value) will find a node with a value matching that of the passed value and return the index of that node, if it doesn't find the matching value it will return null.
It makes a 'count' variable that will keep track of the current index (starting at 0 index), then it enters a while loop that will loop until the current node is null (past the last node in the list) in the loop we evaluate if the current node value is equal to that of the passed value, if so, we return the count, otherwise we set the new current node to the next node and increment the count, then we start the while loop again. If the while loop exits without finding the matching value, the outer scope will return null.