toString() method works. toString() takes the linked list tree and displays it as a string. Example: (HEAD) -> (node value) -> null
It works by making a 'string' variable, that's set to an empty string, this will hold the string we want to ultimately return. Using a while loop that will loop until the current node (starting off at the head node) is equal to null (past the last node), we will add the current node's value as a string to the 'string' variable, followed by assigning the new current node to the next node, then we start the while loop again. Once we exit the while loop, we know that the last item the null value, so we add 'null' to the 'string' variable. Finally, we return the string variable.