removeAt(index) method works. The removeAt(index) method will remove an node at the passed index from the linked list.
The method first assigns the preceding node and the next node of the node at the index passed, their own variables for reference. Next, it makes a variable giving the length of the list that is eligible for removing an element. Then it checks to see if the index is eligible, if the index is 0 (the head node has no previous node and it is not removable) or if the index is larger than the length of the list - then we return an error. Otherwise, we check to see if the index is the length of the list (this means it's the last element in the list, which means we have to assign the previous node's next node to null), else, we assign the previous node's next node to the node following the index.